'''Windows:''' Start of build system port to windows  * Just the Math library right now  * Small tweaks to the math library to obey MSVC warnings  * Tests are disabled on windows  * Updates to bootstrap script to deal with windows specifics (needs linux testing now)

+++ packages/math/src/Math.cpp
@@ -59,11 +59,11 @@ namespace math {
     {
         msAngleUnit = AU_DEGREE;
 
-        mTrigTableSize = trigTableSize;
+        mTrigTableSize = (int)trigTableSize;
         mTrigTableFactor = mTrigTableSize / Math::TWO_PI;
 
-        mSinTable = new Real[mTrigTableSize];
-        mTanTable = new Real[mTrigTableSize];
+        mSinTable = new Real[(unsigned int)mTrigTableSize];
+        mTanTable = new Real[(unsigned int)mTrigTableSize];
 
         buildTrigTables();
     }
